v5.9 - Moved endpoints in routers.
